+++ cdl/tests/errormsgbox_unit.py
@@ -10,7 +10,7 @@ Error message box test
 from qtpy import QtWidgets as QW
 
 from cdl.utils.qthelpers import exec_dialog, qt_app_context
-from cdl.widgets.errormessagebox import ErrorMessageBox
+from cdl.widgets.warningerror import WarningErrorMessageBox
 
 SHOW = True  # Show test in GUI-based test launcher
 
@@ -26,7 +26,7 @@ def test_error_message_box():
         except ValueError:
             context = "Test_error_message_box." * 5
             tip = "This error may occured when testing the error message box. " * 10
-            dlg = ErrorMessageBox(win, context, tip)
+            dlg = WarningErrorMessageBox(win, "error", context, tip)
             exec_dialog(dlg)