+++ cdl/utils/misc.py
@@ -12,9 +12,6 @@ from __future__ import annotations
 import os.path as osp
 import re
 import subprocess
-import warnings
-from collections.abc import Callable
-from typing import Any
 
 import numpy as np
 
@@ -74,10 +71,3 @@ def go_to_error(text: str) -> None:
         args = Conf.console.external_editor_args.get().format(**fdict).split(" ")
         editor_path = Conf.console.external_editor_path.get()
         subprocess.run([editor_path] + args, shell=True, check=False)
-
-
-def wng_func(func: Callable, args: tuple[Any]) -> np.ndarray | tuple[np.ndarray]:
-    """Wrapper function to ignore RuntimeWarning"""
-    with warnings.catch_warnings():
-        warnings.simplefilter("ignore", RuntimeWarning)
-        return func(*args)