+++ cdl/utils/qthelpers.py
@@ -30,7 +30,6 @@ from qtpy import QtWidgets as QW
 from cdl.config import APP_NAME, DATETIME_FORMAT, SHOTPATH, Conf, _, get_old_log_fname
 from cdl.env import execenv
 from cdl.utils.misc import to_string
-from cdl.widgets.errormessagebox import ErrorMessageBox
 
 
 def close_widgets_and_quit(screenshot=False) -> None:
@@ -289,25 +288,6 @@ def qt_try_except(message=None, context=None):
         return method_wrapper
 
     return qt_try_except_decorator
-
-
-@contextmanager
-def qt_try_context(parent: QW.QWidget, context: str = None, tip: str = None):
-    """Try...except Qt widget context manager"""
-    try:
-        yield
-    except Exception:  # pylint: disable=broad-except
-        dlg = ErrorMessageBox(parent, context, tip)
-        exec_dialog(dlg)
-        if execenv.test_mode:
-            #  In test mode, we want to raise the exception because test cases
-            #  are supposed to work without any error. In real life, we want to
-            #  avoid raising the exception because it would stop the application,
-            #  and exceptions could be related to non-critical errors due to
-            #  external libraries.
-            raise
-    finally:
-        pass
 
 
 @contextmanager