+++ cdl/widgets/warningerror.py
@@ -4,7 +4,7 @@
 # Pierre Raybaut
 
 """
-Module providing an error message box
+Module providing a warning/error message box
 """
 
 import traceback
@@ -16,8 +16,9 @@ from guidata.widgets.console.shell import PythonShellWidget
 from qtpy import QtCore as QC
 from qtpy import QtWidgets as QW
 
-from cdl.config import _
+from cdl.config import Conf, _
 from cdl.utils.misc import go_to_error
+from cdl.utils.qthelpers import exec_dialog
 
 
 def insert_spaces(text: str, nbchars: int) -> str:
@@ -50,26 +51,54 @@ def insert_spaces(text: str, nbchars: int) -> str:
     return new_text
 
 
-class ErrorMessageBox(QW.QDialog):
-    """Error message box"""
+class WarningErrorMessageBox(QW.QDialog):
+    """Warning/Error message box
+
+    Args:
+        parent (QW.QWidget): parent widget
+        category (str): message category ("error" or "warning")
+        context (str, optional): context. Defaults to None.
+        message (str, optional): message. Defaults to None.
+        tip (str, optional): tip. Defaults to None.
+    """
 
     def __init__(
-        self, parent: QW.QWidget, context: str = None, tip: str = None
+        self,
+        parent: QW.QWidget,
+        category: str,
+        context: str = None,
+        message: str = None,
+        tip: str = None,
     ) -> None:
         super().__init__(parent)
-        title = parent.window().objectName()
-        self.setWindowTitle(title)
+        assert category in ("error", "warning")
+        self.setWindowTitle(parent.window().objectName())
+
         self.shell = PythonShellWidget(self, read_only=True)
         self.shell.go_to_error.connect(go_to_error)
         font = get_font(CONF, "console")
         font.setPointSize(9)
         self.shell.set_font(font)
-        self.shell.insert_text(traceback.format_exc() * 10, at_end=True, error=True)
+        message = traceback.format_exc() if message is None else message
+        self.shell.insert_text(message, at_end=True, error=True)
 
         bbox = QW.QDialogButtonBox(QW.QDialogButtonBox.Ok)
         bbox.accepted.connect(self.accept)
+        if category == "warning":
+            bbox.addButton(QW.QDialogButtonBox.Ignore).clicked.connect(self.ignore)
 
         layout = QW.QVBoxLayout()
+
+        if category == "error":
+            width, height = 725, 400
+            icon = "MessageBoxCritical"
+            tb_title = _("Error message")
+            tb_text = _("The following traceback may help to understand the problem:")
+        else:
+            width, height = 725, 200
+            icon = "MessageBoxWarning"
+            tb_title = _("Warning message")
+            tb_text = _("Please take into account the following warning message:")
 
         if context is not None:
             context = insert_spaces(context, 80)
@@ -79,7 +108,7 @@ class ErrorMessageBox(QW.QDialog):
             ct_layout = QW.QHBoxLayout()
             ct_image_layout = QW.QVBoxLayout()
             ct_image = QW.QLabel()
-            ct_image.setPixmap(get_std_icon("MessageBoxWarning").pixmap(24, 24))
+            ct_image.setPixmap(get_std_icon(icon).pixmap(24, 24))
             ct_image.setSizePolicy(QW.QSizePolicy.Fixed, QW.QSizePolicy.Fixed)
             ct_image_layout.addWidget(ct_image)
             ct_image_layout.addStretch()
@@ -94,15 +123,11 @@ class ErrorMessageBox(QW.QDialog):
             )
             layout.addWidget(ct_groupbox)
 
-        tb_groupbox = QW.QGroupBox(_("Traceback"), self)
+        tb_groupbox = QW.QGroupBox(tb_title, self)
         tb_layout = QW.QVBoxLayout()
-        tb_text = _("The following traceback may help to understand the problem:")
         tb_layout.addWidget(QW.QLabel(tb_text))
         tb_layout.addWidget(self.shell)
         tb_groupbox.setLayout(tb_layout)
-        tb_groupbox.setSizePolicy(
-            QW.QSizePolicy.MinimumExpanding, QW.QSizePolicy.MinimumExpanding
-        )
         layout.addWidget(tb_groupbox)
 
         if tip is not None:
@@ -126,7 +151,45 @@ class ErrorMessageBox(QW.QDialog):
             layout.addWidget(tip_groupbox)
 
         layout.addSpacing(10)
+        if category == "warning":
+            layout.addWidget(
+                QW.QLabel(
+                    _(
+                        "Please click on the 'Ignore' button to "
+                        "ignore this warning next time."
+                    )
+                )
+            )
+            layout.addSpacing(10)
+
         layout.addWidget(bbox)
 
         self.setLayout(layout)
-        self.resize(800, 500)
+        self.resize(width, height)
+
+    def ignore(self):
+        """Ignore warning next time"""
+        Conf.proc.ignore_warnings.set(True)
+        self.accept()
+
+
+def show_warning_error(
+    parent: QW.QWidget,
+    category: str,
+    context: str = None,
+    message: str = None,
+    tip: str = None,
+) -> None:
+    """Show error message
+
+    Args:
+        parent (QW.QWidget): parent widget
+        category (str): message category ("error" or "warning")
+        context (str, optional): context. Defaults to None.
+        message (str, optional): message. Defaults to None.
+        tip (str, optional): tip. Defaults to None.
+    """
+    if category == "warning" and Conf.proc.ignore_warnings.get():
+        return
+    dlg = WarningErrorMessageBox(parent, category, context, message, tip)
+    exec_dialog(dlg)